code/bot/serverless/front-slash-handler/requirements.txt
PyNaCl
requests
boto3